requirements.txt
--find-links https://download.pytorch.org/whl/cu117/torch_stable.html
torch==2.0.1+cu117 ; sys_platform == 'linux' or sys_platform == 'win32'
torch==2.0.1 ; sys_platform == 'darwin'
ctranslate2==3.20.0
stanza==1.1.1
sentencepiece==0.1.99
requests==2.31.0
PyYAML==6.0.1
sacrebleu==2.3.1
subword-nmt>=0.3.7
OpenNMT-py==3.4.1
tensorboard==2.14.0
six==1.16.0
iso639==0.1.4
sacremoses==0.0.53
removedup==1.0.6
fastshuffle==1.0.1
unbabel-comet==2.2.2
urllib3~=2.2.3
torch~=2.0.1+cu117
numpy~=1.26.4
tqdm~=4.66.5
jsonargparse~=3.13.1
transformers~=4.44.2
langchain~=0.3.0
mlflow~=2.16.2